Avoid attempting duplicate secure URLs
Various Autodiscover methods can resolve to the same URL. If an
incorrect password has been provided then these will all fail regardless
of them being the correct endpoint.

To fail as quickly as possible in this case we record the URLs as they
are attempted in order to avoid attempting to use them again at a later
stage.